Created 'RotatingRectangle' + Cleaned up Collider.
Created rotating rectangle, this extends libgdx rectangle class but allows us to give it an angle field, which we then utilise in collider to work out the rectangles hitbox.

Also cleaned up collider, adding a Polygon return to 'getRectangleHitbox' as this is how we will represent the rectangles hitboxes. Also removed 'CircleOnCircle' method as this is obsolete since there will be no circle on circle collisions.